# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Skill {
  id: string;
  name: string;
  description: string;
  symbol: string;
  accent: string;
  examples: string[];
}

export type QuestionActivityType = typeof QuestionActivityType[keyof typeof QuestionActivityType];


export const QuestionActivityType = {
  standard: 'standard',
  visual: 'visual',
} as const;

export interface Question {
  id: string;
  skillId: string;
  prompt: string;
  answer: number;
  level: number;
  hint: string;
  activityType?: QuestionActivityType;
}

export interface AttemptInput {
  questionId: string;
  skillId: string;
  answer: number;
  expectedAnswer: number;
  level: number;
  /** @nullable */
  explanation?: string | null;
}

export interface AttemptResult {
  correct: boolean;
  feedback: string;
  explanation: string;
  nextLevel: number;
  mastery: number;
  streak?: number;
  recommendation?: string;
}

export interface LearningAnalysisInput {
  /**
     * @minimum 1
     * @maximum 6
     */
  gradeLevel: number;
  skillId: string;
  skill: string;
  question: string;
  correctAnswer: number;
  studentAnswer: number;
  /**
     * @minLength 1
     * @maxLength 500
     */
  studentExplanation: string;
  /** @minimum 0 */
  previousAttemptCount: number;
  previousWasIncorrect: boolean;
  /**
     * @minimum 1
     * @maximum 5
     */
  currentDifficulty: number;
}

export type LearningAnalysisResultConfidence = typeof LearningAnalysisResultConfidence[keyof typeof LearningAnalysisResultConfidence];


export const LearningAnalysisResultConfidence = {
  low: 'low',
  medium: 'medium',
  high: 'high',
} as const;

export type LearningAnalysisResultRecommendedNextActivity = typeof LearningAnalysisResultRecommendedNextActivity[keyof typeof LearningAnalysisResultRecommendedNextActivity];


export const LearningAnalysisResultRecommendedNextActivity = {
  retry: 'retry',
  simpler: 'simpler',
  visual: 'visual',
  similar: 'similar',
} as const;

export interface LearningAnalysisResult {
  explanation: string;
  misconception: string;
  confidence: LearningAnalysisResultConfidence;
  recommendedStrategy: string;
  recommendedNextActivity: LearningAnalysisResultRecommendedNextActivity;
}

export interface Mastery {
  skillId: string;
  attempted: number;
  correct: number;
  mastery: number;
  level: number;
  streak: number;
}

export interface Activity {
  id: string;
  skillId: string;
  skillName: string;
  correct: boolean;
  question: string;
  createdAt: string;
}

export type GetPracticeQuestionParams = {
/**
 * @minimum 1
 * @maximum 5
 */
level?: number;
/**
 * @minimum 0
 */
seed?: number;
activity?: GetPracticeQuestionActivity;
};

export type GetPracticeQuestionActivity = typeof GetPracticeQuestionActivity[keyof typeof GetPracticeQuestionActivity];


export const GetPracticeQuestionActivity = {
  standard: 'standard',
  visual: 'visual',
} as const;