# lib/api-zod/src/generated/types/activity.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Activity {
  id: string;
  skillId: string;
  skillName: string;
  correct: boolean;
  question: string;
  createdAt: string;
}