# lib/api-zod/src/generated/types/attemptInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AttemptInput {
  questionId: string;
  skillId: string;
  answer: number;
  expectedAnswer: number;
  level: number;
  /** @nullable */
  explanation?: string | null;
}