# lib/api-zod/src/generated/types/attemptResult.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AttemptResult {
  correct: boolean;
  feedback: string;
  explanation: string;
  nextLevel: number;
  mastery: number;
  streak?: number;
  recommendation?: string;
}