# lib/api-zod/src/generated/types/learningAnalysisInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface LearningAnalysisInput {
  /**
     * @minimum 1
     * @maximum 6
     */
  gradeLevel: number;
  skillId: string;
  skill: string;
  question: string;
  correctAnswer: number;
  studentAnswer: number;
  /**
     * @minLength 1
     * @maxLength 500
     */
  studentExplanation: string;
  /** @minimum 0 */
  previousAttemptCount: number;
  previousWasIncorrect: boolean;
  /**
     * @minimum 1
     * @maximum 5
     */
  currentDifficulty: number;
}