# lib/api-zod/src/generated/types/learningAnalysisResult.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { LearningAnalysisResultConfidence } from './learningAnalysisResultConfidence';
import type { LearningAnalysisResultRecommendedNextActivity } from './learningAnalysisResultRecommendedNextActivity';

export interface LearningAnalysisResult {
  explanation: string;
  misconception: string;
  confidence: LearningAnalysisResultConfidence;
  recommendedStrategy: string;
  recommendedNextActivity: LearningAnalysisResultRecommendedNextActivity;
}