# lib/api-zod/src/generated/types/mastery.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Mastery {
  skillId: string;
  attempted: number;
  correct: number;
  mastery: number;
  level: number;
  streak: number;
}