# lib/api-zod/src/generated/types/question.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { QuestionActivityType } from './questionActivityType';

export interface Question {
  id: string;
  skillId: string;
  prompt: string;
  answer: number;
  level: number;
  hint: string;
  activityType?: QuestionActivityType;
}