# lib/api-zod/src/generated/types/skill.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Skill {
  id: string;
  name: string;
  description: string;
  symbol: string;
  accent: string;
  examples: string[];
}